Add files via upload
Java project to determine based on data files of monitoring internet traffic of an IP camera whether or not there was movement in a scene. Data files contain different video codecs for public IP cameras, no sensitive information is present.
The program takes in one data file at a time, which is located in a variable in code and determines whether or not movement occurred.